s13_langchain/requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-classic>=1.0.0
openai>=1.0.0,<2.0.0
python-dotenv>=1.0.0
